+++ plugins/TripleOscillator/TripleOscillator.h
@@ -78,11 +78,11 @@ private:
 	float m_volumeRight;
 
 	// normalized detuning -> x/sampleRate
-	float m_detuningLeft;
-	float m_detuningRight;
+	double m_detuningLeft;
+	double m_detuningRight;
 	// normalized offset -> x/360
-	float m_phaseOffsetLeft;
-	float m_phaseOffsetRight;
+	double m_phaseOffsetLeft;
+	double m_phaseOffsetRight;
 	bool m_useWaveTable;
 
 	friend class TripleOscillator;
