+++ src/core/Oscillator.cpp
@@ -56,7 +56,7 @@ Oscillator::Oscillator(const IntModel *wave_shape_model,
 			const IntModel *mod_algo_model,
 			const float &freq,
 			const float &detuning_div_samplerate,
-			const float &phase_offset,
+			const double &phase_offset,
 			const float &volume,
 			Oscillator *sub_osc) :
 	m_waveShapeModel(wave_shape_model),
@@ -535,7 +535,7 @@ void Oscillator::updateFM( sampleFrame * _ab, const fpp_t _frames,
 // should be called every time phase-offset is changed...
 inline void Oscillator::recalcPhase()
 {
-	if( !typeInfo<float>::isEqual( m_phaseOffset, m_ext_phaseOffset ) )
+	if( !typeInfo<double>::isEqual( m_phaseOffset, m_ext_phaseOffset ) )
 	{
 		m_phase -= m_phaseOffset;
 		m_phaseOffset = m_ext_phaseOffset;
@@ -549,10 +549,10 @@ inline void Oscillator::recalcPhase()
 
 inline bool Oscillator::syncOk( float _osc_coeff )
 {
-	const float v1 = m_phase;
+	const double v1 = m_phase;
 	m_phase += _osc_coeff;
 	// check whether m_phase is in next period
-	return( floorf( m_phase ) > floorf( v1 ) );
+	return( floor( m_phase ) > floor( v1 ) );
 }
